Initial commit of alignment code
Re-factored camera-alignment-core code from microscopy-foc-post-processing repo and placed it into camera_alignment_core.py.
Added 2 functions:
similarity_matrix_transform: takes alignment matrix and a slice of an image and applies the matrix transform on the slice, outputting the transformed slice

align_image:
Takes the full "CZYX" image, the channel_to_index map for that image, and the alignment matrix, and aligns all channels that need to be aligned. Outputs a full "CZYX" array that has all aligned data (not cropped)